+++ src/app/modules/about-us/index.styles.tsx
@@ -3,15 +3,18 @@ import {
   ModuleBackground,
   ModuleContainer,
 } from "@/app/components/shared/containers/index.styles";
-import Image from "next/image";
 
-const AboutUsModuleContainer = styled(ModuleContainer)`
-  display: flex;
+type MobileProps = {
+  mobile?: boolean | null;
+};
+
+const AboutUsModuleContainer = styled(ModuleContainer)<MobileProps>`
+  display: ${(props) => (props.mobile ? "block" : "flex")};
   justify-content: space-between;
   align-items: center;
-  padding: 2rem;
+  padding: ${(props) => (props.mobile ? "0.5rem" : "2rem")};
   position: relative;
-  margin: 0 auto;
+  margin: 0 auto; 
 `;
 
 const AboutUsModuleBackground = styled(ModuleBackground)`
@@ -25,90 +28,4 @@ const AboutUsModuleBackground = styled(ModuleBackground)`
   background-position: center;
 `;
 
-const LeftContainer = styled.div`
-  max-width: 50vw;
-  padding-right: 3vw;
-  padding-left: 2vw;
-  position: relative;
-
-  h1 {
-    letter-spacing: 0px;
-    font-size: 60px;
-    font-weight: 700;
-    margin-bottom: 1.5rem;
-    color: #fff;
-    text-align: left;
-    line-height: 60px;
-    font-style: normal;
-
-    .highlight {
-      position: relative;
-      display: inline-block;
-    }
-  }
-`;
-
-const AboutGraphic = styled(Image)`
-  position: absolute;
-  top: -15px;
-  left: 400px;
-  transform: translateY(-30%);
-`;
-
-const AboutUsDescription = styled.div`
-  background-color: #fff;
-  border-radius: 8px;
-  padding: 1rem 2rem;
-  margin-bottom: 2rem;
-  font-size: 16px;
-  text-align: left;
-  letter-spacing: 0px;
-`;
-
-const ImagesContainer = styled.div`
-  display: grid;
-  grid-template-columns: 1fr;
-  grid-template-rows: auto auto;
-  position: relative;
-  padding-right: 2vw;
-
-  & > :nth-child(1) {
-    grid-column: 1 / 2;
-    align-self: end;
-  }
-
-  & > :nth-child(2) {
-    grid-column: 2 / 3;
-  }
-
-  & > :nth-child(3) {
-    grid-column: 1 / 3;
-    justify-self: center;
-    margin-left: -40px;
-  }
-  & > :nth-child(4) {
-    position: absolute;
-    bottom: 5%;
-    left: 100%;
-    margin-left: -110px;
-    align-self: center;
-  }
-`;
-
-const StyledImage = styled(Image)`
-  display: block;
-  border: 10px solid #fff;
-  border-radius: 4px;
-  object-fit: cover;
-  margin: 8px;
-`;
-
-export {
-  AboutUsModuleContainer,
-  AboutUsModuleBackground,
-  LeftContainer,
-  AboutUsDescription,
-  AboutGraphic,
-  ImagesContainer,
-  StyledImage,
-};
+export {AboutUsModuleBackground, AboutUsModuleContainer};
