+++ src/app/modules/about-us/index.tsx
@@ -1,17 +1,14 @@
-import {
-  AboutUsModuleBackground,
-  AboutUsModuleContainer,
-  LeftContainer,
-  AboutUsDescription,
-  AboutGraphic,
-  ImagesContainer,
-  StyledImage,
-} from "@/app/modules/about-us/index.styles";
+import AboutUsDesktop from "@/app/components/about-us/desktop/index";
+import AboutUsMobile from "@/app/components/about-us/mobile/index";
+import { AboutUsModuleBackground, AboutUsModuleContainer } from "@/app/modules/about-us/index.styles";
+import { useMobileDetect } from "@/app/hooks/useMobileDetect";
 import Manrope from "@/app/components/shared/fonts/manrope";
 import Image from "next/image";
 import SpaceGrotesk from "@/app/components/shared/fonts/space-grotesk";
 
 const AboutUs = () => {
+  const isMobile = useMobileDetect();
+
   const aboutUsDescription = (
     <>
       <p>
@@ -29,52 +26,13 @@ const AboutUs = () => {
   );
 
   return (
-    <AboutUsModuleContainer>
-      <AboutUsModuleBackground src="/background/about-us.svg" />
-      <LeftContainer>
-        <SpaceGrotesk>
-          <h1>
-            Canada’s <span className="highlight">first</span> <br /> student-run
-            hackathon
-          </h1>
-        </SpaceGrotesk>
-        <AboutGraphic
-          src="/assets/about/Graphic.svg"
-          alt="Gradient Graphic"
-          width={75}
-          height={75}
-        />
-        <Manrope>
-          <AboutUsDescription>{aboutUsDescription}</AboutUsDescription>
-        </Manrope>
-      </LeftContainer>
-
-      <ImagesContainer>
-        <StyledImage
-          src="/assets/about/about-1.svg"
-          alt="UofT Hacks executive speaking to a crowd"
-          width={160}
-          height={160}
-        />
-        <StyledImage
-          src="/assets/about/about-2.svg"
-          alt="UofT Hacks sponsors"
-          width={350}
-          height={200}
-        />
-        <StyledImage
-          src="/assets/about/about-3.svg"
-          alt="UofT Hacks executive speaking to a crowd"
-          width={400}
-          height={200}
-        />
-        <Image
-          src="/assets/about/Keys.svg"
-          alt="Keys Icon"
-          width={75}
-          height={170}
-        />
-      </ImagesContainer>
+    <AboutUsModuleContainer mobile={isMobile}>
+    <AboutUsModuleBackground src="/background/about-us.svg" />
+    {isMobile ? (
+        <AboutUsMobile/>
+        ) : (
+          <AboutUsDesktop/>
+    )}
     </AboutUsModuleContainer>
   );
 };
